requirements.txt
Django==2.2.1
djangorestframework==3.9.4
huey==2.0.1
redis==3.2.1
requests==2.22.0
